start/requirements.txt
azure-functions
requests
pillow
onnxruntime
